+++ packages/ChangelogLinker/src/Worker/DiffLinksToVersionsWorker.php
@@ -5,7 +5,7 @@ namespace Symplify\ChangelogLinker\Worker;
 use Nette\Utils\Strings;
 use Symplify\ChangelogLinker\Contract\Worker\WorkerInterface;
 
-final class CompleteDiffLinksToVersionsWorker implements WorkerInterface
+final class DiffLinksToVersionsWorker implements WorkerInterface
 {
     /**
      * @var string[]
