+++ packages/ChangelogLinker/tests/ChangelogApplicationSource/after/01.md
@@ -5,3 +5,4 @@
 ### Added
 
 - [#123] some text
+
