+++ packages/ChangelogLinker/tests/ChangelogApplicationSource/after/03.md
@@ -5,3 +5,5 @@
 ### Added
 
 - [#505] some existing pull-request
+
+[#505]: https://github.com/Symplify/Symplify/pull/505
