+++ packages/ChangelogLinker/tests/Worker/DiffLinksToVersionsWorker/Source/after/01.md
@@ -1,0 +1,7 @@
+# Changelog
+
+## [v2.0.0]
+
+## [v1.0.0]
+
+[v2.0.0]: https://github.com/Symplify/Symplify/compare/v1.0.0...v2.0.0
