+++ packages/ChangelogLinker/tests/Worker/DiffLinksToVersionsWorker/Source/before/01.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## [v2.0.0]
+
+## [v1.0.0]
